% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/warpgroupTools.R
\name{eicMatFromList}
\alias{eicMatFromList}
\title{A helper function to interpolate a list of EIC's into a rectangular array with optional down/up-sampling and smoothing}
\usage{
eicMatFromList(eic.l, eic.resample.target = Inf, upsample.force = F,
  smooth.n = 1, normalize = T)
}
\arguments{
\item{eic.l}{A list of matrices respresnting EICs.  Each matrix has rows "rt", "intensity", and "scan"}

\item{eic.resample.target}{Numeric.  If less than one the resulting EICs will be of length max*length.target.  If greater than 1 resulting EICs will be of length length.target.  If Inf resulting EICs will be of length max.}

\item{upsample.force}{Boolean.  Must be true for resulting EICs to be of length > max.}

\item{smooth.n}{Integer. The number of points to include in the moving average.}

\item{normalize}{Boolean. If T all EICs will be normalized to 1.0}
}
\value{
An array containing each EIC's resulting intensity, retention time, and scan.
}
\description{
\code{eicMatFromList} returns an array containing each EIC's resulting intensity, retention time, and scan.
}
\details{
This function takes a list of EIC's (intensities, retention times, and indices for each scan) and interpolates them such that they are the same length and normalizes the intensities to 1.
}

