% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xcmsWarpgroup.R
\name{group.warpgroup}
\alias{group.warpgroup}
\title{A wrapper for supplying xcmsSet objects to Warpgroup.}
\usage{
group.warpgroup(xs, xr.l, sc.max.drift, ppm.max.drift, sc.aligned.lim,
  output.groups = F, sc.aligned.factor = 1, detailed.groupinfo = F,
  min.peaks = 1)
}
\arguments{
\item{xs}{An xcmsSet object with initial grouping information.}

\item{xr.l}{A list contanining xcmsRaw objects for each sample in the xcmsSet object in order.}

\item{sc.max.drift}{Integer.  The maximum time drift expected for a peak in the data set in scans.  Used when looking for missing peaks.}

\item{ppm.max.drift}{Integer. The maximum mass drift expected for a peak in the data set in ppm.  Used when looking for missing peaks.}

\item{sc.aligned.lim}{Integer. Peak bounds after alignment are considered the same if they are within this limit.}

\item{output.groups}{Boolean. If \code{TRUE} the output is a list of warpgroup outputs for every group rather than an xcmsSet.  Allows for better integration parameter selection with \code{\link{warpgroupsToXs}}.}

\item{sc.aligned.factor}{Float. Experimental feature where graph edges are weighted proportionally to the distance between the aligned peak bounds. Higher numbers emphasize closer peak bounds.}

\item{detailed.groupinfo}{Boolean. Returns several extra descriptors of the warping and graph clustering.}

\item{min.peaks}{Integer. Groups with fewer peaks than \code{min.peaks} are skipped.}
}
\value{
An xcmsSet with groups and peaks replaced by warpgroup generated ones.
}
\description{
\code{group.warpgroup} returns an xcmsSet with the original groups and peaks replaced by those returned by the warpgroup algorithm.
}
\details{
This function simply iterates over each group in the xcmsSet and runs \code{\link{warpgroup}} on them. The resulting warpgroup output is then re-integrated and returned as a new xcmsSet.
The resulting xcmsSet is ready for a call to \code{\link{diffreport}} and does not need to be grouped.  The results will be redundant and should be filtered appropriately.
}
\seealso{
See \url{https://github.com/nathaniel-mahieu/warpgroup} for examples.
}

