% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/warpgroup.R
\name{warpgroup}
\alias{warpgroup}
\title{Warpgroup a set of peak bounds.}
\usage{
warpgroup(ps, eic.mat.s, sc.aligned.lim, sc.aligned.factor = 1,
  detailed.groupinfo = F, min.peaks = 1, tw = "dtw", pct.pad = 0.1)
}
\arguments{
\item{ps}{Matrix. Rows representing peak regions detected in \code{eic.mat.s} and containing columns \code{sc, scmin, scmax, sample}.}

\item{eic.mat.s}{Matrix. Chromatograms, one per sample. Columns corresponding to samples and rows corresponding to time with values of intensity.}

\item{sc.aligned.lim}{Integer. Peak bounds after alignment are considered the same if they are within this limit.}

\item{sc.aligned.factor}{Float. Experimental feature where graph edges are weighted proportionally to the distance between the aligned peak bounds. Higher numbers emphasize closer peak bounds.}

\item{detailed.groupinfo}{Logical. Returns several extra descriptors of the warping and graph clustering.}

\item{min.peaks}{Integer. Supplied groups with fewer than this number of peak bounds will be skipped.}

\item{tw}{Character. \code{"tw"} or \code{"dtw"} Allows the selection of the desired correspondance algorithm}

\item{pct.pad}{Float. Defines how much padding is added to the ends of each chromatogram.}
}
\value{
A list, one entry for each distinct chromatographic region detected.  Within each entry is a matrix, with one row for each sample describing the start and end of the corresponding peak region as indices of the supplied \code{eic.mat.s}.
}
\description{
\code{warpgroup} returns a list of all distinct chromatographic regions determined from the supplied peak bounds for every sample.
}
\details{
This function implements the time-warping based chromatogram correspondance determination, splitting of supplied peak bounds into distinct groups, and selection of the correct region to integrate for missing values.
The warpgroup approach can be applied to a variety of situations but is most effective when supplied groups are of high quality.

Warpgroup first determines the correspondance between supplied EICs using a time warping technique (by default dynamic time warping).  Based on this correspondance warpgroup splits the supplied peak bounds into sub groups, each which describe similar chromatographic regions.  Warpgroup then determines the appropriate integration region for samples missing peaks.
}
\seealso{
See \url{https://github.com/nathaniel-mahieu/warpgroup} for examples.
}

