% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/xcmsIntegrate.R
\name{warpgroupsToXs}
\alias{warpgroupsToXs}
\title{A wrapper for integration of warpgroups and packaging as an xcmsSet object.}
\usage{
warpgroupsToXs(xs, groups, xr.l, ppm.padding = 0.1, min.ppm.width = 0)
}
\arguments{
\item{xs}{An xcmsSet object.}

\item{groups}{The list of warpgroups as output by \code{group.warpgroup(output.groups=TRUE)}}

\item{xr.l}{A list contanining xcmsRaw objects for each sample in the xcmsSet object in order.}

\item{ppm.padding}{Float. An expansion of the mass window for integration.}

\item{min.ppm.width}{Float. Any mass windows narrower than this are expanded to this width.}
}
\value{
An xcmsSet with groups and peaks replaced by warpgroup generated ones.
}
\description{
\code{warpgroupsToXs} returns an xcmsSet with the original groups and peaks replaced by those described in \code{groups}.
}
\details{
This function simply iterates over each group, integrates the appropriate region from xr.l and packages the results as an xcmsSet.
}
\seealso{
See \url{https://github.com/nathaniel-mahieu/warpgroup} for examples.
}

